Frontend handles feedbacks for multiple services
The frontend has been upgraded to handle feedbacks belonging to different services. A service is identified by a public token the frontend must have in order to call the welcome page. Through the public token, a custom message is fetched from the backend and presented to the user. Once the rating is submitted, it carries also the public token, in order to identify the service the rating belongs to.

ToDo: upgrade the backend to handle this new behaviour.