Welcome message handled
The backend exposes an action to provide the welcomeMessage to the frontend. It selects the message by means of the public token returned by the service creation action.

Note that, when a great number of service is managed by the application, the public token field should be indexed on the database in order to avoid a full table scan.